# Retro-Forum-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HiddenFreeways API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /states/{slug}:
    get:
      operationId: getStateBySlug
      tags: [states]
      summary: Get a state by slug with its locations and pinned threads
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: State detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ------------------- CATEGORIES -------------------
...
  /messages:
    get:
      operationId: listPrivateChats
      tags: [dms]
      summary: List the current user's private chats
      responses:
        "200":
          description: DM threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrivateChatSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: openPrivateChat
      tags: [dms]
      summary: Open or create a private chat with another user (by username)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenPrivateChatBody"
      responses:
        "200":
          description: Existing chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateChatSummary"
        "201":
          description: Created chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateChatSummary"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/{id}:
    get:
      operationId: listPrivateChatMessages
      tags: [dms]
      summary: Get messages in a private chat
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: sinceId
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomMessagesResponse"
        "403":
          description: Not a participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: sendPrivateChatMessage
      tags: [dms]
      summary: Send a private message in a chat
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomMessage"
        "403":
          description: Not a participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ---------- AUTH ----------
...
    State:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        abbreviation:
          type: string
        centerLat:
          type: number
        centerLng:
          type: number
        zoom:
          type: integer
      required: [id, slug, name, abbreviation, centerLat, centerLng, zoom]
...
    StateDetail:
      type: object
      properties:
        state:
          $ref: "#/components/schemas/State"
        locations:
          type: array
          items:
            $ref: "#/components/schemas/Location"
        pinnedThreads:
          type: array
          items:
            $ref: "#/components/schemas/ThreadSummary"
      required: [state, locations, pinnedThreads]

    # ---------- CATEGORIES ----------
...
    Location:
      type: object
      properties:
        id:
          type: integer
        stateId:
          type: integer
        stateSlug:
          type: string
        stateName:
          type: string
        name:
          type: string
        description:
          type: string
        city:
          type: ["string", "null"]
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [active, demolished, sealed, watched]
        risk:
          type: string
          enum: [low, medium, high, extreme]
        createdById:
          type: integer
        createdByUsername:
          type: string
        threadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - stateId
        - stateSlug
        - stateName
        - name
        - description
        - city
        - latitude
        - longitude
        - status
        - risk
        - createdById
        - createdByUsername
        - threadCount
        - createdAt
...
    ThreadSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: string
        categoryId:
          type: integer
        categorySlug:
          type: string
        categoryName:
          type: string
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        stateSlug:
          type: ["string", "null"]
        authorId:
          type: integer
        authorUsername:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        isPinned:
          type: boolean
        isLocked:
          type: boolean
        replyCount:
          type: integer
        lastActivityAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - excerpt
        - categoryId
        - categorySlug
        - categoryName
        - locationId
        - locationName
        - stateSlug
        - authorId
        - authorUsername
        - authorAvatarUrl
        - isPinned
        - isLocked
        - replyCount
        - lastActivityAt
        - createdAt
...
    RoomMessage:
      type: object
      properties:
        id:
          type: integer
        body:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorTrustLevel:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - body
        - authorId
        - authorUsername
        - authorTrustLevel
        - createdAt
    RoomMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/RoomMessage"
      required: [messages]
    SendChatMessageBody:
      type: object
      properties:
        body:
          type: string
      required: [body]
...
    PrivateChatSummary:
      type: object
      properties:
        id:
          type: integer
        otherUserId:
          type: integer
        otherUsername:
          type: string
        otherTrustLevel:
          type: integer
        lastMessageBody:
          type: ["string", "null"]
        lastActivityAt:
          type: string
          format: date-time
      required:
        - id
        - otherUserId
        - otherUsername
        - otherTrustLevel
        - lastMessageBody
        - lastActivityAt
    OpenPrivateChatBody:
      type: object
      properties:
        username:
          type: string
      required: [username]